[ERM-5124] RE: CR
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Dev/Features/EFCF/BLCore;C25964

+++ Source/ApplicationServices/2Gis.Erm.BLCore.Operations/Concrete/Old/Integration/OneC/ExportAccountDetailsToServiceBusForBranchHandler.cs
@@ -156,7 +156,7 @@ namespace DoubleGis.Erm.BLCore.Operations.Concrete.Old.Integration.OneC
                     Amount = GetAccountDetailAmount(x),
                     SignedOnDate = x.OrderSignupDateUtc,
                     ClientOrderNumber = x.ClientOrderNumber,
-                    OrderType = x.OrderType,
+                    OrderType = (int)x.OrderType,
                     OrderNumber = x.OrderNumber,
                     MediaInfo = x.ElectronicMedia,
                     LegalEntityBranchCode1C = x.BranchOfficeOrganizationUnitSyncCode1C,
